--- server/llm/interfaces.go
@@ -1,44 +1,0 @@
-package llm
-
-import (
-	"image"
-	"io"
-
-	"github.com/mattermost/mattermost-plugin-ai/server/llm/subtitles"
-)
-
-type LLMConfig struct {
-	Model              string
-	MaxGeneratedTokens int
-	EnableVision       bool
-}
-
-type LanguageModelOption func(*LLMConfig)
-
-func WithModel(model string) LanguageModelOption {
-	return func(cfg *LLMConfig) {
-		cfg.Model = model
-	}
-}
-
-func WithMaxGeneratedTokens(maxGeneratedTokens int) LanguageModelOption {
-	return func(cfg *LLMConfig) {
-		cfg.MaxGeneratedTokens = maxGeneratedTokens
-	}
-}
-
-type LanguageModel interface {
-	ChatCompletion(conversation BotConversation, opts ...LanguageModelOption) (*TextStreamResult, error)
-	ChatCompletionNoStream(conversation BotConversation, opts ...LanguageModelOption) (string, error)
-
-	CountTokens(text string) int
-	TokenLimit() int
-}
-
-type Transcriber interface {
-	Transcribe(file io.Reader) (*subtitles.Subtitles, error)
-}
-
-type ImageGenerator interface {
-	GenerateImage(prompt string) (image.Image, error)
-}
